Feat/nuevas entidades

+++ chemiconsult/src/main/java/com/chemiconsult/repository/UserRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<UserDE,Integer> {
+public interface UserRepository extends JpaRepository<UserDE,Long> {
 
     Optional<UserDE> findByUsername(String username);
